Move ControllerRequestClient from pinot-controller to pinot-common, so other modules is able to use it to call controller
