+++ pinot-common/src/test/java/org/apache/pinot/util/http/MultiGetRequestTest.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.pinot.common.http;
+package org.apache.pinot.util.http;
 
 import com.google.common.base.Throwables;
 import com.sun.net.httpserver.HttpExchange;
@@ -34,6 +34,7 @@ import java.util.concurrent.ExecutionException;
 import java.util.concurrent.Executors;
 import org.apache.commons.httpclient.MultiThreadedHttpConnectionManager;
 import org.apache.commons.httpclient.methods.GetMethod;
+import org.apache.pinot.common.utils.http.MultiGetRequest;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.testng.Assert;
