+++ pinot-controller/src/main/java/org/apache/pinot/controller/util/CompletionServiceHelper.java
@@ -28,7 +28,7 @@ import java.util.concurrent.Executor;
 import org.apache.commons.httpclient.HttpConnectionManager;
 import org.apache.commons.httpclient.URI;
 import org.apache.commons.httpclient.methods.GetMethod;
-import org.apache.pinot.common.http.MultiGetRequest;
+import org.apache.pinot.common.utils.http.MultiGetRequest;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
